requirements.txt
tqdm
pyyaml

sentencepiece
datasets

torch
torchdata
torchtext
tensorboard